Replace usages of WifiHelper with WifiConnector
