+++ skunkworks_crow/src/main/java/org/odk/share/activities/SendActivity.java
@@ -19,10 +19,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import org.odk.share.R;
-import org.odk.share.controller.WifiHelper;
 import org.odk.share.controller.WifiHotspotHelper;
 import org.odk.share.events.HotspotEvent;
 import org.odk.share.events.UploadEvent;
+import org.odk.share.network.WifiConnector;
 import org.odk.share.rx.RxEventBus;
 import org.odk.share.rx.schedulers.BaseSchedulerProvider;
 import org.odk.share.services.HotspotService;
@@ -106,9 +106,10 @@ public class SendActivity extends InjectableActivity {
             Timber.e("Port not available for socket communication");
             finish();
         }
-        WifiHelper wifiHelper = new WifiHelper(this);
-        if (wifiHelper.getWifiManager().isWifiEnabled()) {
-            wifiHelper.disableWifi(null);
+
+        WifiConnector wifiConnector = new WifiConnector(this);
+        if (wifiConnector.isWifiEnabled()) {
+            wifiConnector.disableWifi();
         }
 
         isHotspotInitiated = false;
@@ -174,7 +175,7 @@ public class SendActivity extends InjectableActivity {
             Intent intent = new Intent(getApplicationContext(), HotspotService.class);
             intent.setAction(HotspotService.ACTION_STATUS);
             startService(intent);
-            
+
             currentConfig = wifiHotspot.getCurrConfig();
             startSending();
         }
@@ -267,7 +268,7 @@ public class SendActivity extends InjectableActivity {
             intent.setAction(HotspotService.ACTION_STATUS);
             startService(intent);
             Timber.d("Started hotspot N");
-            currentConfig = wifiHotspot.getCurrConfig();            
+            currentConfig = wifiHotspot.getCurrConfig();
             startSending();
         }
     }
