+++ skunkworks_crow/src/main/java/org/odk/share/network/WifiBroadcastReceiver.java
@@ -5,20 +5,22 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.NetworkInfo;
 import android.net.wifi.WifiManager;
+import android.support.annotation.Nullable;
 
 import timber.log.Timber;
 
 public class WifiBroadcastReceiver extends BroadcastReceiver {
 
+    @Nullable
     private WifiBroadcastListener listener;
 
-    public WifiBroadcastReceiver(WifiBroadcastListener listener) {
+    public WifiBroadcastReceiver(@Nullable WifiBroadcastListener listener) {
         this.listener = listener;
     }
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        if (intent.getAction() != null) {
+        if (intent.getAction() != null && listener != null) {
             switch (intent.getAction()) {
                 case WifiManager.NETWORK_STATE_CHANGED_ACTION:
                     NetworkInfo networkInfo = intent.getParcelableExtra(WifiManager.EXTRA_NETWORK_INFO);
